+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 ### Added
 
 * User and site admin management capabilities for user email addresses are improved.
+* Enable new search timeout behavior by enabling `experimentalFeatures.searchTimeoutParameterEnabled` in your site config.
+  * Adds a new `timeout:` parameter to customize the timeout for searches. It defaults to 10s and may not be set higher than 1m.
+  * The value of the `timeout:` parameter is a string that can be parsed by [time.Duration](https://golang.org/pkg/time/#ParseDuration) (e.g. "100ms", "2s").
+  * When `timeout:` is not provided, search optimizes for retuning results as soon as possible and will include slower kinds of results (e.g. symbols) only if they are found quickly.
+  * When `timeout:` is provided, all result kinds are given the full timeout to complete.
 
 ## 2.6.8